Address review: shared object metadata resolution and isDefined guards
- Extract getObjectMetadataItemBySingularNameOrThrow, shared by the chat
  target placement and the record index side panel opener
- Use isDefined for the nullable state guards on the record index panel
  page, and fix its formatting
